+++ @navikt/aksel-stylelint/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @navikt/aksel-stylelint
 
+## 8.9.0
+
 ## 8.8.0
 
 ### Patch Changes